Move imported/exported dependencies
