+++ package.json
@@ -38,7 +38,7 @@
     "ajv": "^8.11.0",
     "axios": "^1.3.4",
     "bignumber.js": "9.0.2",
-    "ethereumjs-abi": "github:ProjectWyvern/ethereumjs-abi",
+    "ethereumjs-abi": "git+https://github.com/ProjectWyvern/ethereumjs-abi.git",
     "ethereumjs-util": "^5.2.0",
     "ethers": "^5.6.6",
     "fbemitter": "^3.0.0",
