+++ src/utils/schemas/ERC721/index.ts
@@ -1,9 +1,10 @@
+import { Schema } from "../schema";
 import {
+  AbiType,
+  StateMutability,
   FunctionInputKind,
   FunctionOutputKind,
-  StateMutability,
-} from "../../../types";
-import { AbiType, Schema } from "../schema";
+} from "../types";
 
 interface NonFungibleContractType {
   id: string;
