+++ src/utils/schemas/rinkeby/testRinkebyNFT/index.ts
@@ -1,11 +1,11 @@
+import { Schema } from "../../schema";
 import {
   AbiType,
+  EventInputKind,
   FunctionInputKind,
   FunctionOutputKind,
   StateMutability,
-} from "../../../../types";
-import { Schema } from "../../schema";
-import { EventInputKind } from "../../types";
+} from "../../types";
 
 type TestRinkebyNFTType = string;
